student-app/backend/requirements.txt
# Core ML frameworks
transformers==4.53.0
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
accelerate==1.1.0
numpy<2.0
timm

# MLX framework (Apple Silicon optimized - optional)
# Install only if you plan to use MLX models
# mlx-vlm>=0.3.2

# File processing and utilities
watchdog==5.0.3
python-dotenv==1.0.0
protobuf>=3.20.0
requests>=2.31.0

# Web API
fastapi==0.104.1
uvicorn==0.24.0
python-multipart>=0.0.20